create hokusai registry command group
